moved tree to rctree
